# PatientService/apifox_openapi.yaml
openapi: 3.0.3
info:
  title: HGY Hospital - PatientService (Waiting & Extra Apply)
  version: 1.0.0
servers:
  - url: "{{baseUrl}}"
    description: "Set environment variable baseUrl to http://localhost:8082"
paths:
  /api/registrations/waiting:
    post:
      summary: Create a waiting (候补挂号)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWaitingRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultWaitingDto'
    get:
      summary: Get waiting list by schedule
      parameters:
        - in: query
          name: scheduleRecordId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
    delete:
      summary: Cancel waiting
      parameters:
        - in: query
          name: patientId
          schema:
            type: string
          required: true
        - in: query
          name: waitingId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Success
  /api/registrations/waiting/patient:
    get:
      summary: Get waiting by patient and optional date
      parameters:
        - in: query
          name: patientId
          schema:
            type: string
          required: true
        - in: query
          name: date
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Success
  /api/registrations/waiting/confirm:
    post:
      summary: Confirm waiting -> convert to registration (系统/后台调用)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                waitingId:
                  type: string
      responses:
        '200':
          description: Success

  /api/extra-apply:
    post:
      summary: Submit an extra-apply (患者申请加号)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientExtraApplyCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultPatientExtraApply'
    get:
      summary: List extra-apply for a patient
      parameters:
        - in: query
          name: patientId
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: OK
  /api/extra-apply/{id}:
    get:
      summary: Get extra-apply by id
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: OK
  /api/extra-apply/{id}/approve:
    put:
      summary: Approve an extra-apply
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
        - in: query
          name: approverId
          schema:
            type: integer
      responses:
        '200':
          description: OK
  /api/extra-apply/{id}/reject:
    put:
      summary: Reject an extra-apply
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
        - in: query
          name: approverId
          schema:
            type: integer
        - in: query
          name: rejectReason
          schema:
            type: string
      responses:
        '200':
          description: OK

components:
  schemas:
    Result:
      type: object
      properties:
        code:
          type: integer
        msg:
          type: string
        data:
          type: object
    CreateWaitingRequest:
      type: object
      required:
        - patientId
        - scheduleRecordId
      properties:
        patientId:
          type: string
        scheduleRecordId:
          type: string
    WaitingDto:
      type: object
      properties:
        waitingId:
          type: string
        patientId:
          type: string
        scheduleRecordId:
          type: string
        applyTime:
          type: string
          format: date-time
        status:
          type: string
        position:
          type: integer
        limitCount:
          type: integer
    ResultWaitingDto:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/WaitingDto'
    PatientExtraApplyCreate:
      type: object
      required:
        - patientId
        - departmentId
        - doctorId
        - appointmentDate
        - reason
      properties:
        patientId:
          type: integer
        departmentId:
          type: integer
        doctorId:
          type: integer
        appointmentDate:
          type: string
          format: date
        reason:
          type: string
    PatientExtraApply:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        departmentId:
          type: integer
        doctorId:
          type: integer
        appointmentDate:
          type: string
          format: date
        reason:
          type: string
        status:
          type: string
        locked:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        rejectReason:
          type: string
    ResultPatientExtraApply:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PatientExtraApply'